// src/types/ProductFilters.ts
export type ProductFilters = {
  ram: string[];
  capacity: string[];
  color: string[];
  size: string[];
  matrixType: string[];
  priceRange: [number, number];
};

export type FilterOptions = {
  ram: string[];
  capacity: string[];
  color: string[];
  size: string[];
  matrixType: string[];
  priceRange: [number, number];
};